requirements.txt
# Core FastAPI and server
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1

# AI and ML
google-generativeai==0.7.2
pinecone-client==3.2.2
openai==1.107.0

# Database and ORM
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Caching and Performance
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2

# Authentication and Security (Python 3.9 compatible versions)
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart>=0.0.7
cryptography==3.4.8
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# Validation and Serialization
pydantic==2.8.2
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring and Logging
structlog==23.2.0
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0

# Rate Limiting and Middleware
slowapi==0.1.9

# Background Tasks
celery==5.3.4

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development Tools
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Optional integrations
streamlit==1.37.1
requests==2.32.3
slack_bolt==1.19.0
slack_sdk==3.33.1
aiohttp==3.9.1
pandas==2.2.2

# BM25 and Text Processing for Hybrid Search
rank-bm25==0.2.2
nltk==3.8.1
scikit-learn==1.3.2

# Health checks and utilities
psutil==5.9.6
python-dateutil==2.8.2
